refactor: split generate page script
