+++ src/admin/ts/generate/generate-page-handlers.ts
@@ -1,328 +1,24 @@
-import {
-  nuclenShowElement,
-  nuclenHideElement,
-  nuclenUpdateProgressBarStep,
-  nuclenCheckCreditsAjax,
-  nuclenToggleSummaryFields,
-} from './generate-page-utils';
-import {
-  NuclenStartGeneration,
-  NuclenPollAndPullUpdates,
-  nuclenFetchWithRetry,
-} from '../nuclen-admin-generate';
+import { getGeneratePageElements } from './elements';
+import { initStep1 } from './step1';
+import { initStep2 } from './step2';
+import { initGoBack, initRestart, initSummaryToggle } from './navigation';
 
 export function initGeneratePage(): void {
-  const step1 = document.getElementById('nuclen-step-1') as HTMLDivElement | null;
-  const step2 = document.getElementById('nuclen-step-2') as HTMLDivElement | null;
-  const updatesSection = document.getElementById('nuclen-updates-section') as HTMLDivElement | null;
-  const updatesContent = document.getElementById('nuclen-updates-content') as HTMLDivElement | null;
-  const restartBtn = document.getElementById('nuclen-restart-btn') as HTMLButtonElement | null;
-  const getPostsBtn = document.getElementById('nuclen-get-posts-btn') as HTMLButtonElement | null;
-  const goBackBtn = document.getElementById('nuclen-go-back-btn') as HTMLButtonElement | null;
-  const generateForm = document.getElementById('nuclen-generate-form') as HTMLFormElement | null;
-  const submitBtn = document.getElementById('nuclen-submit-btn') as HTMLButtonElement | null;
-  const postsCountEl = document.getElementById('nuclen-posts-count') as HTMLSpanElement | null;
-  const stepBar1 = document.getElementById('nuclen-step-bar-1');
-  const stepBar2 = document.getElementById('nuclen-step-bar-2');
-  const stepBar3 = document.getElementById('nuclen-step-bar-3');
-  const stepBar4 = document.getElementById('nuclen-step-bar-4');
-  const creditsInfoEl = document.getElementById('nuclen-credits-info') as HTMLParagraphElement | null;
+  const elements = getGeneratePageElements();
 
-  const NUCLEN_REST_RECEIVE_CONTENT_ENDPOINT =
-    (window as any).nuclenAdminVars?.rest_receive_content ||
-    '/wp-json/nuclear-engagement/v1/receive-content';
-  const restNonce = (window as any).nuclenAdminVars?.rest_nonce || '';
+  // Initial UI setup
+  elements.step1 && elements.step1.classList.remove('nuclen-hidden');
+  elements.step2 && elements.step2.classList.add('nuclen-hidden');
+  elements.updatesSection && elements.updatesSection.classList.add('nuclen-hidden');
+  elements.restartBtn && elements.restartBtn.classList.add('nuclen-hidden');
+  elements.stepBar1 && elements.stepBar1.classList.add('nuclen-step-current');
+  elements.stepBar2 && elements.stepBar2.classList.add('nuclen-step-todo');
+  elements.stepBar3 && elements.stepBar3.classList.add('nuclen-step-todo');
+  elements.stepBar4 && elements.stepBar4.classList.add('nuclen-step-todo');
 
-  // Initial UI
-  nuclenShowElement(step1);
-  nuclenHideElement(step2);
-  nuclenHideElement(updatesSection);
-  nuclenHideElement(restartBtn);
-  nuclenUpdateProgressBarStep(stepBar1, 'current');
-  nuclenUpdateProgressBarStep(stepBar2, 'todo');
-  nuclenUpdateProgressBarStep(stepBar3, 'todo');
-  nuclenUpdateProgressBarStep(stepBar4, 'todo');
-
-  // Step 1 => Get Posts
-  getPostsBtn?.addEventListener('click', () => {
-    if (!(window as any).nuclenAjax || !(window as any).nuclenAjax.ajax_url) {
-      alert('Error: Ajax is not configured properly. Please check the plugin settings.');
-      return;
-    }
-    if (postsCountEl) {
-      postsCountEl.innerText = 'Loading posts...';
-    }
-    const formData = new FormData();
-    formData.append('action', 'nuclen_get_posts_count');
-    if ((window as any).nuclenAjax?.nonce) {
-      formData.append('security', (window as any).nuclenAjax.nonce);
-    }
-    const postStatusEl = document.getElementById('nuclen_post_status') as HTMLSelectElement | null;
-    const categoryEl = document.getElementById('nuclen_category') as HTMLSelectElement | null;
-    const authorEl = document.getElementById('nuclen_author') as HTMLSelectElement | null;
-    const postTypeEl = document.getElementById('nuclen_post_type') as HTMLSelectElement | null;
-    const workflowEl = document.getElementById('nuclen_generate_workflow') as HTMLSelectElement | null;
-    const allowRegenEl = document.getElementById('nuclen_allow_regenerate_data') as HTMLInputElement | null;
-    const protectRegenEl = document.getElementById('nuclen_regenerate_protected_data') as HTMLInputElement | null;
-    formData.append('nuclen_post_status', postStatusEl ? postStatusEl.value : '');
-    formData.append('nuclen_category', categoryEl ? categoryEl.value : '');
-    formData.append('nuclen_author', authorEl ? authorEl.value : '');
-    formData.append('nuclen_post_type', postTypeEl ? postTypeEl.value : '');
-    formData.append('nuclen_generate_workflow', workflowEl ? workflowEl.value : '');
-    formData.append('nuclen_allow_regenerate_data', allowRegenEl && allowRegenEl.checked ? '1' : '0');
-    formData.append('nuclen_regenerate_protected_data', protectRegenEl && protectRegenEl.checked ? '1' : '0');
-
-    nuclenFetchWithRetry((window as any).nuclenAjax.ajax_url || '', {
-      method: 'POST',
-      body: formData,
-      credentials: 'same-origin',
-    })
-      .then((r) => r.json())
-      .then(async (data) => {
-        if (!data.success) {
-          if (postsCountEl) {
-            postsCountEl.innerText = 'Error retrieving post count.';
-          }
-          if (data.data?.message) {
-            if (data.data.message.includes('Invalid API key')) {
-              alert('Your Gold Code (API key) is invalid. Please create a new one on the NE app and enter it on the plugin Setup page.');
-            } else if (data.data.message.includes('Invalid WP App Password')) {
-              alert('Your WP App Password is invalid. Please go to the plugin Setup page and re-generate it.');
-            } else {
-              alert(data.data.message);
-            }
-          }
-          return;
-        }
-        const count = data.data.count as number;
-        const foundPosts = data.data.post_ids;
-        const selectedPostIdsEl = document.getElementById('nuclen_selected_post_ids') as HTMLInputElement | null;
-        if (selectedPostIdsEl) {
-          selectedPostIdsEl.value = JSON.stringify(foundPosts);
-        }
-        const workflowEl2 = document.getElementById('nuclen_generate_workflow') as HTMLSelectElement | null;
-        const selectedWorkflowEl = document.getElementById('nuclen_selected_generate_workflow') as HTMLInputElement | null;
-        if (selectedWorkflowEl && workflowEl2) {
-          selectedWorkflowEl.value = workflowEl2.value;
-        }
-        const summaryFormatElStep1 = document.getElementById('nuclen_summary_format') as HTMLSelectElement | null;
-        const summaryLengthElStep1 = document.getElementById('nuclen_summary_length') as HTMLSelectElement | null;
-        const summaryNumberElStep1 = document.getElementById('nuclen_summary_number_of_items') as HTMLSelectElement | null;
-        const selectedSummaryFormatEl = document.getElementById('nuclen_selected_summary_format') as HTMLInputElement | null;
-        const selectedSummaryLengthEl = document.getElementById('nuclen_selected_summary_length') as HTMLInputElement | null;
-        const selectedSummaryNumberEl = document.getElementById('nuclen_selected_summary_number_of_items') as HTMLInputElement | null;
-        if (summaryFormatElStep1 && selectedSummaryFormatEl) {
-          selectedSummaryFormatEl.value = summaryFormatElStep1.value;
-        }
-        if (summaryLengthElStep1 && selectedSummaryLengthEl) {
-          selectedSummaryLengthEl.value = summaryLengthElStep1.value;
-        }
-        if (summaryNumberElStep1 && selectedSummaryNumberEl) {
-          selectedSummaryNumberEl.value = summaryNumberElStep1.value;
-        }
-        const selectedPostStatusEl = document.getElementById('nuclen_selected_post_status') as HTMLInputElement | null;
-        if (selectedPostStatusEl && postStatusEl) {
-          selectedPostStatusEl.value = postStatusEl.value;
-        }
-        const selectedPostTypeEl = document.getElementById('nuclen_selected_post_type') as HTMLInputElement | null;
-        if (selectedPostTypeEl && postTypeEl) {
-          selectedPostTypeEl.value = postTypeEl.value;
-        }
-        nuclenHideElement(step1);
-        nuclenShowElement(step2);
-        nuclenUpdateProgressBarStep(stepBar1, 'done');
-        nuclenUpdateProgressBarStep(stepBar2, 'current');
-        if (count === 0) {
-          if (postsCountEl) {
-            postsCountEl.innerText = 'No posts found with these filters.';
-          }
-          if (submitBtn) {
-            nuclenHideElement(submitBtn);
-          }
-          return;
-        }
-        if (postsCountEl) {
-          postsCountEl.innerText = `Number of posts to process: ${count}`;
-        }
-        try {
-          const remainingCredits = await nuclenCheckCreditsAjax();
-          const neededCredits = count;
-          if (creditsInfoEl) {
-            creditsInfoEl.textContent = `This will consume ${neededCredits} credit(s). You have ${remainingCredits} left.`;
-          }
-          if (remainingCredits < neededCredits) {
-            alert('Not enough credits. Please top up or reduce the number of posts.');
-            if (submitBtn) {
-              submitBtn.disabled = true;
-            }
-          } else if (submitBtn) {
-            nuclenShowElement(submitBtn);
-            submitBtn.disabled = false;
-          }
-        } catch (err: any) {
-          console.error('Error fetching remaining credits:', err);
-          if (creditsInfoEl) {
-            creditsInfoEl.textContent = `Unable to retrieve your credits: ${err.message}`;
-          }
-          if (submitBtn) {
-            submitBtn.disabled = false;
-          }
-        }
-      })
-      .catch((error) => {
-        console.error('Error retrieving post count:', error);
-        if (postsCountEl) {
-          postsCountEl.innerText = 'Error retrieving post count. Please try again later.';
-        }
-      });
-  });
-
-  // Step 2 -> Go Back
-  goBackBtn?.addEventListener('click', () => {
-    nuclenHideElement(step2);
-    nuclenShowElement(step1);
-    if (postsCountEl) {
-      postsCountEl.innerText = '';
-    }
-    if (creditsInfoEl) {
-      creditsInfoEl.textContent = '';
-    }
-    nuclenUpdateProgressBarStep(stepBar1, 'current');
-    nuclenUpdateProgressBarStep(stepBar2, 'todo');
-  });
-
-  // Step 2 -> Generate
-  generateForm?.addEventListener('submit', async (event) => {
-    event.preventDefault();
-    if (!(window as any).nuclenAdminVars || !(window as any).nuclenAdminVars.ajax_url) {
-      alert('Error: WP Ajax config not found. Please check the plugin settings.');
-      return;
-    }
-    nuclenUpdateProgressBarStep(stepBar2, 'done');
-    nuclenUpdateProgressBarStep(stepBar3, 'current');
-    nuclenShowElement(updatesSection);
-    if (updatesContent) {
-      updatesContent.innerText = 'Processing posts... Do NOT leave this page until the process is complete.';
-    }
-    nuclenHideElement(step2);
-    if (submitBtn) {
-      submitBtn.disabled = true;
-    }
-    try {
-      const formDataObj = Object.fromEntries(new FormData(generateForm).entries());
-      const startResp = await NuclenStartGeneration(formDataObj);
-      const generationId =
-        startResp.data?.generation_id ||
-        startResp.generation_id ||
-        'gen_' + Math.random().toString(36).substring(2);
-      NuclenPollAndPullUpdates({
-        intervalMs: 5000,
-        generationId,
-        onProgress: (processed, total) => {
-          const safeProcessed = processed === undefined ? 0 : processed;
-          const safeTotal = total === undefined ? '' : total;
-          if (updatesContent) {
-            updatesContent.innerText = `Processed ${safeProcessed} of ${safeTotal} posts so far...`;
-          }
-        },
-        onComplete: async ({ failCount, finalReport, results, workflow }) => {
-          nuclenUpdateProgressBarStep(stepBar3, 'done');
-          nuclenUpdateProgressBarStep(stepBar4, 'current');
-          if (results && typeof results === 'object') {
-            try {
-              const payload = { workflow, results };
-              const storeResp = await fetch(NUCLEN_REST_RECEIVE_CONTENT_ENDPOINT, {
-                method: 'POST',
-                headers: {
-                  'Content-Type': 'application/json',
-                  'X-WP-Nonce': restNonce,
-                },
-                credentials: 'include',
-                body: JSON.stringify(payload),
-              });
-              const storeData = await storeResp.json();
-              if (storeResp.ok && !storeData.code) {
-                console.log('Bulk content stored in WP meta successfully:', storeData);
-              } else {
-                console.error('Error storing bulk content in WP meta:', storeData);
-              }
-            } catch (err) {
-              console.error('Error storing bulk content in WP meta:', err);
-            }
-          }
-          if (updatesContent) {
-            if (failCount && finalReport) {
-              updatesContent.innerText = `Some posts failed. ${finalReport.message || ''}`;
-              nuclenUpdateProgressBarStep(stepBar4, 'failed');
-            } else {
-              updatesContent.innerText = 'All posts processed successfully! Your content has been saved.';
-              nuclenUpdateProgressBarStep(stepBar4, 'done');
-            }
-          }
-          if (submitBtn) {
-            submitBtn.disabled = false;
-          }
-          nuclenShowElement(restartBtn);
-        },
-        onError: (errMsg: string) => {
-          nuclenUpdateProgressBarStep(stepBar3, 'failed');
-          if (errMsg.includes('Invalid API key')) {
-            alert('Your API key is invalid. Please go to the Setup page and enter a new one.');
-          } else if (errMsg.includes('Invalid WP App Password')) {
-            alert('Your WP App Password is invalid. Please re-generate it on the Setup page.');
-          } else if (errMsg.includes('Not enough credits')) {
-            alert('Not enough credits. Please top up your account or reduce the number of posts.');
-          } else {
-            alert(`Error: ${errMsg}`);
-          }
-          if (updatesContent) {
-            updatesContent.innerText = `Error: ${errMsg}`;
-          }
-          if (submitBtn) {
-            submitBtn.disabled = false;
-          }
-          nuclenShowElement(restartBtn);
-        },
-      });
-    } catch (error: any) {
-      nuclenUpdateProgressBarStep(stepBar3, 'failed');
-      if (error.message.includes('Invalid API key')) {
-        alert('Your API key is invalid. Please go to the Setup page and enter a new one.');
-      } else if (error.message.includes('Invalid WP App Password')) {
-        alert('Your WP App Password is invalid. Please go to the Setup page and re-generate it.');
-      } else if (error.message.includes('Not enough credits')) {
-        alert('Not enough credits. Please top up or reduce posts.');
-      } else {
-        alert(`Error starting generation: ${error.message}`);
-      }
-      if (submitBtn) {
-        submitBtn.disabled = false;
-      }
-      nuclenShowElement(restartBtn);
-    }
-  });
-
-  nuclenToggleSummaryFields();
-  const generateTypeEl = document.getElementById('nuclen_generate_workflow') as HTMLSelectElement | null;
-  const summaryFormatEl = document.getElementById('nuclen_summary_format') as HTMLSelectElement | null;
-  generateTypeEl?.addEventListener('change', nuclenToggleSummaryFields);
-  summaryFormatEl?.addEventListener('change', nuclenToggleSummaryFields);
-
-  // Restart
-  restartBtn?.addEventListener('click', () => {
-    nuclenHideElement(updatesSection);
-    nuclenHideElement(restartBtn);
-    nuclenHideElement(step2);
-    if (postsCountEl) {
-      postsCountEl.innerText = '';
-    }
-    if (creditsInfoEl) {
-      creditsInfoEl.textContent = '';
-    }
-    nuclenShowElement(step1);
-    nuclenUpdateProgressBarStep(stepBar1, 'current');
-    nuclenUpdateProgressBarStep(stepBar2, 'todo');
-    nuclenUpdateProgressBarStep(stepBar3, 'todo');
-    nuclenUpdateProgressBarStep(stepBar4, 'todo');
-  });
+  initStep1(elements);
+  initStep2(elements);
+  initGoBack(elements);
+  initRestart(elements);
+  initSummaryToggle();
 }
